lib: serial: cast 'void *' to 'unsigned long'
Cast 'void *' to 'unsigned long' to avoid [-Werror=int-conversion]
error when passing a pointer address to a function that expects a
plain integer address.

+++ lib/utils/serial/sifive-uart.c
@@ -112,7 +112,7 @@ int sifive_uart_init(void *base, u32 in_freq, u32 baudrate)
 
 	sbi_console_set_device(&sifive_console);
 
-	return sbi_domain_root_add_memrange(base, PAGE_SIZE, PAGE_SIZE,
+	return sbi_domain_root_add_memrange((unsigned long)base, PAGE_SIZE, PAGE_SIZE,
 					    (SBI_DOMAIN_MEMREGION_MMIO |
 					    SBI_DOMAIN_MEMREGION_SHARED_SURW_MRW));
 }
